Move functions for resolving locations
