+++ extensions/ql-vscode/src/language-support/ast-viewer/ast-viewer.ts
@@ -17,7 +17,7 @@ import { basename } from "path";
 
 import { DatabaseItem } from "../../databases/local-databases";
 import { UrlValue, BqrsId } from "../../pure/bqrs-cli-types";
-import { showLocation } from "../../interface-utils";
+import { showLocation } from "../../databases/local-databases/locations";
 import {
   isStringLoc,
   isWholeFileLoc,
