+++ extensions/ql-vscode/test/vscode-tests/no-workspace/interface-utils.test.ts
@@ -1,20 +1,8 @@
-import {
-  Uri,
-  Location,
-  Range,
-  Position,
-  window,
-  ViewColumn,
-  WebviewPanel,
-} from "vscode";
+import { Uri, ViewColumn, WebviewPanel, window } from "vscode";
 import { basename } from "path";
-import { fileSync, FileResult } from "tmp";
-import {
-  fileUriToWebviewUri,
-  tryResolveLocation,
-} from "../../../src/interface-utils";
+import { FileResult, fileSync } from "tmp";
+import { fileUriToWebviewUri } from "../../../src/interface-utils";
 import { getDefaultResultSetName } from "../../../src/pure/interface-types";
-import { mockDatabaseItem } from "../utils/mocking.helpers";
 
 describe("interface-utils", () => {
   describe("webview uri conversion", () => {
@@ -79,96 +67,6 @@ describe("interface-utils", () => {
       expect(getDefaultResultSetName(["a", "b", "#select"])).toBe("#select");
       expect(getDefaultResultSetName(["a", "b"])).toBe("a");
       expect(getDefaultResultSetName([])).toBeUndefined();
-    });
-  });
-
-  describe("resolveWholeFileLocation", () => {
-    it("should resolve a whole file location", () => {
-      const databaseItem = mockDatabaseItem();
-      expect(
-        tryResolveLocation("file://hucairz:0:0:0:0", databaseItem),
-      ).toEqual(new Location(Uri.file("abc"), new Range(0, 0, 0, 0)));
-    });
-
-    it("should resolve a five-part location edge case", () => {
-      const databaseItem = mockDatabaseItem();
-      expect(
-        tryResolveLocation("file://hucairz:1:1:1:1", databaseItem),
-      ).toEqual(new Location(Uri.file("abc"), new Range(0, 0, 0, 1)));
-    });
-
-    it("should resolve a five-part location", () => {
-      const databaseItem = mockDatabaseItem();
-
-      expect(
-        tryResolveLocation(
-          {
-            startColumn: 1,
-            endColumn: 3,
-            startLine: 4,
-            endLine: 5,
-            uri: "hucairz",
-          },
-          databaseItem,
-        ),
-      ).toEqual(
-        new Location(
-          Uri.parse("abc"),
-          new Range(new Position(4, 3), new Position(3, 0)),
-        ),
-      );
-      expect(databaseItem.resolveSourceFile).toHaveBeenCalledTimes(1);
-      expect(databaseItem.resolveSourceFile).toHaveBeenCalledWith("hucairz");
-    });
-
-    it("should resolve a five-part location with an empty path", () => {
-      const databaseItem = mockDatabaseItem();
-
-      expect(
-        tryResolveLocation(
-          {
-            startColumn: 1,
-            endColumn: 3,
-            startLine: 4,
-            endLine: 5,
-            uri: "",
-          },
-          databaseItem,
-        ),
-      ).toBeUndefined();
-    });
-
-    it("should resolve a string location for whole file", () => {
-      const databaseItem = mockDatabaseItem();
-
-      expect(
-        tryResolveLocation("file://hucairz:0:0:0:0", databaseItem),
-      ).toEqual(new Location(Uri.parse("abc"), new Range(0, 0, 0, 0)));
-      expect(databaseItem.resolveSourceFile).toHaveBeenCalledTimes(1);
-      expect(databaseItem.resolveSourceFile).toHaveBeenCalledWith("hucairz");
-    });
-
-    it("should resolve a string location for five-part location", () => {
-      const databaseItem = mockDatabaseItem();
-
-      expect(
-        tryResolveLocation("file://hucairz:5:4:3:2", databaseItem),
-      ).toEqual(
-        new Location(
-          Uri.parse("abc"),
-          new Range(new Position(4, 3), new Position(2, 2)),
-        ),
-      );
-      expect(databaseItem.resolveSourceFile).toHaveBeenCalledTimes(1);
-      expect(databaseItem.resolveSourceFile).toHaveBeenCalledWith("hucairz");
-    });
-
-    it("should resolve a string location for invalid string", () => {
-      const databaseItem = mockDatabaseItem();
-
-      expect(
-        tryResolveLocation("file://hucairz:x:y:z:a", databaseItem),
-      ).toBeUndefined();
     });
   });
 });
